requirements.txt
tweepy>=4.0
neo4j
textblob
colorama
python-dotenv
openai